requirements_pubnub.txt
# Requirements for PubNub-based teleoperation
pubnub>=10.4.1
pyserial>=3.5
scservo_sdk  # Install from Feetech SDK
msgpack>=1.0.0  # For efficient data serialization
colorama>=0.4.6  # For colored terminal output 